pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backlog-pilot"
version = "0.1.0"
description = "A Python CLI tool that integrates with Azure DevOps and GitHub to help Roo Code burn down your backlog"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Burndown Agent Team"}
]
keywords = ["azure-devops", "github", "cli", "backlog", "automation", "roo-code"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.1.1",
    "azure-devops>=7.1.0b4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/aselunar/Burndown-Agent"
Repository = "https://github.com/aselunar/Burndown-Agent"
Issues = "https://github.com/aselunar/Burndown-Agent/issues"

[project.scripts]
backlog-pilot = "backlog_pilot.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
backlog_pilot = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=backlog_pilot --cov-report=term-missing"
